Time: 56 ms (66.50%) | Memory: 108.1 MB (63.50%) - LeetSync

+++ 4242-sum-of-gcd-of-formed-pairs/sum-of-gcd-of-formed-pairs.java
@@ -10,7 +10,7 @@ class Solution {
         Arrays.sort(prefixGcd);
         int j = prefixGcd.length-1;
         for(int i=0; i<prefixGcd.length/2; i++){
-            sum += (long) gcd(prefixGcd[i],prefixGcd[j]);
+            sum += gcd(prefixGcd[i],prefixGcd[j]);
             j--; 
         }
         return sum;
